+++ src/compiler.c
@@ -15441,6 +15441,7 @@ void diamond_program_init_fresh(DiamondProgram *program) {
         [DIAMOND_CLASS_NO_METHOD_ERROR]={"NoMethodError",DIAMOND_CLASS_STANDARD_ERROR},
         [DIAMOND_CLASS_JSON_ERROR]={"JSONError",DIAMOND_CLASS_STANDARD_ERROR},
         [DIAMOND_CLASS_SUPERVISOR_ERROR]={"SupervisorError",DIAMOND_CLASS_STANDARD_ERROR},
+        [DIAMOND_CLASS_SANDBOX_ERROR]={"SandboxError",DIAMOND_CLASS_STANDARD_ERROR},
     };
     program->range_class_index=UINT8_MAX;
     program->class_count=DIAMOND_BUILTIN_CLASS_COUNT;